Added calculate.php (incomplete)
Calculate.php contains a script that divides numbers and formats the
calculation as you would do it when dividing manually. The formatting is
still off and the script only performs the first iteration of the
calculation. Added an include() in index.php to load calculate.php when
two numbers are entered and the button is pressed